requirements.txt
absl-py==0.11.0
aiohttp==3.7.3
appdirs==1.4.4
asgiref==3.3.1
astunparse==1.6.3
async-timeout==3.0.1
attrs==20.3.0
backcall==0.2.0
black==20.8b1
bokeh==2.2.3
cachetools==4.2.1
cchardet==2.1.7
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
cycler==0.10.0
decorator==4.4.2
Django==3.1.7
dominate==2.6.0
fastprogress==1.0.0
filelock==3.0.12
fsspec==0.8.5
future==0.18.2
gast==0.3.3
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.35.0
h5py==2.10.0
idna==2.10
image==1.5.33
imageio==2.9.0
importlib-metadata==3.4.0
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.18.0
jieba==0.42.1
Jinja2==2.11.2
joblib==1.0.0
jsonpatch==1.28
jsonpointer==2.0
Keras==2.4.3
keras-bert==0.86.0
keras-embed-sim==0.8.0
keras-layer-normalization==0.14.0
keras-multi-head==0.27.0
keras-pos-embd==0.11.0
keras-position-wise-feed-forward==0.6.0
Keras-Preprocessing==1.1.2
keras-self-attention==0.46.0
keras-transformer==0.38.0
kiwisolver==1.3.1
ktrain==0.21.3
langdetect==1.0.8
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.3
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
multidict==5.1.0
mypy-extensions==0.4.3
networkx==2.5
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1603479632437/work
oauthlib==3.1.0
olefile==0.46
opencv-contrib-python==4.5.1.48
opt-einsum==3.3.0
packaging==20.8
pandas==1.1.5
parso==0.8.1
pathspec==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow @ file:///tmp/build/80754af9/pillow_1609786740831/work
prompt-toolkit==3.0.14
protobuf==3.14.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.7.4
pyparsing==2.4.7
python-dateutil==2.8.1
pytorch-lightning==1.2.0
pytorch-ssim==0.1
pytz==2020.5
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==22.0.3
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
sacremoses==0.0.43
scikit-image==0.16.2
scikit-learn==0.23.2
scipy==1.6.0
seqeval==1.2.2
six @ file:///tmp/build/80754af9/six_1605205313296/work
sqlparse==0.4.1
syntok==1.3.1
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.1
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
test-tube==0.7.5
threadpoolctl==2.1.0
tokenizers==0.9.4
toml==0.10.2
torch==1.5.0
torch-summary==1.4.5
torchfile==0.1.0
torchsummary==1.5.1
torchvision==0.6.0
tornado==6.1
tqdm==4.56.0
traitlets==5.0.5
transformers==4.2.2
typed-ast==1.4.2
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1598376058250/work
urllib3==1.26.2
visdom==0.1.8.9
wcwidth==0.2.5
websocket-client==0.57.0
Werkzeug==1.0.1
Whoosh==2.7.4
wrapt==1.12.1
yarl==1.6.3
zipp==3.4.0
